+++ ui/src/organizations/components/CreateVariableOverlay.tsx
@@ -1,132 +1,39 @@
 // Libraries
-import React, {PureComponent, ChangeEvent} from 'react'
+import React, {PureComponent} from 'react'
 
 // Styles
 import 'src/organizations/components/CreateVariableOverlay.scss'
 
 // Components
-import {
-  Form,
-  OverlayBody,
-  OverlayHeading,
-  OverlayContainer,
-  Input,
-  OverlayFooter,
-} from 'src/clockface'
-import {Button} from '@influxdata/clockface'
-import FluxEditor from 'src/shared/components/FluxEditor'
+import {OverlayBody, OverlayHeading, OverlayContainer} from 'src/clockface'
+import VariableForm from 'src/organizations/components/VariableForm'
 
 // Types
 import {Variable} from '@influxdata/influx'
-import {
-  ComponentColor,
-  ComponentStatus,
-  ButtonType,
-} from '@influxdata/clockface'
 
 interface Props {
   onCreateVariable: (variable: Variable) => void
-  onCloseModal: () => void
+  onHideOverlay: () => void
   orgID: string
+  initialScript?: string
 }
 
-interface State {
-  name: string
-  script: string
-  nameInputStatus: ComponentStatus
-  errorMessage: string
-}
-
-export default class CreateOrgOverlay extends PureComponent<Props, State> {
-  constructor(props) {
-    super(props)
-    this.state = {
-      name: '',
-      script: '',
-      nameInputStatus: ComponentStatus.Default,
-      errorMessage: '',
-    }
-  }
-
+export default class CreateVariableOverlay extends PureComponent<Props> {
   public render() {
-    const {onCloseModal} = this.props
-    const {nameInputStatus, name, script} = this.state
+    const {onHideOverlay, onCreateVariable, orgID, initialScript} = this.props
 
     return (
       <OverlayContainer maxWidth={1000}>
-        <OverlayHeading
-          title="Create Variable"
-          onDismiss={this.props.onCloseModal}
-        />
-
-        <Form onSubmit={this.handleSubmit}>
-          <OverlayBody>
-            <div className="overlay-flux-editor--spacing">
-              <Form.Element label="Name">
-                <Input
-                  placeholder="Give your variable a name"
-                  name="name"
-                  autoFocus={true}
-                  value={name}
-                  onChange={this.handleChangeInput}
-                  status={nameInputStatus}
-                />
-              </Form.Element>
-            </div>
-
-            <Form.Element label="Value">
-              <div className="overlay-flux-editor">
-                <FluxEditor
-                  script={script}
-                  onChangeScript={this.handleChangeScript}
-                  visibility="visible"
-                  suggestions={[]}
-                />
-              </div>
-            </Form.Element>
-
-            <OverlayFooter>
-              <Button
-                text="Cancel"
-                color={ComponentColor.Danger}
-                onClick={onCloseModal}
-              />
-              <Button
-                text="Create"
-                type={ButtonType.Submit}
-                color={ComponentColor.Primary}
-              />
-            </OverlayFooter>
-          </OverlayBody>
-        </Form>
+        <OverlayHeading title="Create Variable" onDismiss={onHideOverlay} />
+        <OverlayBody>
+          <VariableForm
+            onCreateVariable={onCreateVariable}
+            onHideOverlay={onHideOverlay}
+            orgID={orgID}
+            initialScript={initialScript}
+          />
+        </OverlayBody>
       </OverlayContainer>
     )
-  }
-
-  private handleSubmit = (): void => {
-    const {onCreateVariable, orgID, onCloseModal} = this.props
-
-    onCreateVariable({
-      name: this.state.name,
-      orgID,
-      arguments: {
-        type: 'query',
-        values: {query: this.state.script, language: 'flux'},
-      },
-    })
-
-    onCloseModal()
-  }
-
-  private handleChangeInput = (e: ChangeEvent<HTMLInputElement>) => {
-    const {value, name} = e.target
-
-    const newState = {...this.state}
-    newState[name] = value
-    this.setState(newState)
-  }
-
-  private handleChangeScript = (script: string): void => {
-    this.setState({script})
   }
 }